+++ client/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ## [v0.4.0 Unreleased]
 
+- Upload data source is only enabled to users with the role 'admin' [LANDGRIF-1125](https://vizzuality.atlassian.net/browse/LANDGRIF-1125)
+
 - New hook to check user role and permissions [LANDGRIF-1123](https://vizzuality.atlassian.net/browse/LANDGRIF-1123)
 
 ## [v0.3.5]